GPS-16 - clean up App.vue
