+++ src/components/pages/Main.vue
@@ -1,14 +1,9 @@
 <script setup>
 import { defineProps, defineEmits } from 'vue'
-import GuineaPig from '../guinea-pig/GuineaPig.vue'
 import Cage from '../cage/Cage.vue'
 import StatusMarquee from '../cage/StatusMarquee.vue'
 import CageItemManager from '../cage/CageItemManager.vue'
-import Panel from '../shared/Panel.vue'
-import Needs from '../guinea-pig/Needs.vue'
-import Inventory from '../inventory/Inventory.vue'
-import CageStatus from '../cage/CageStatus.vue'
-import Market from '../market/Market.vue'
+import IconSidebarPanels from '../panels/IconSidebarPanels.vue'
 import { useGuineaPigStore } from '../../stores/guineaPig'
 
 const guineaPigStore = useGuineaPigStore()
@@ -41,45 +36,18 @@ const emit = defineEmits(['closeInventory', 'closeGuineaPig', 'closeNeeds', 'clo
     </div>
 
     <!-- Modal Panels -->
-    <Panel 
-      :isOpen="showInventory" 
-      title="Your Inventory" 
-      @close="emit('closeInventory')"
-    >
-      <Inventory />
-    </Panel>
-
-    <Panel 
-      :isOpen="showGuineaPig" 
-      title="Guinea Pig Information" 
-      @close="emit('closeGuineaPig')"
-    >
-      <GuineaPig />
-    </Panel>
-
-    <Panel 
-      :isOpen="showNeeds" 
-      title="Guinea Pig Needs" 
-      @close="emit('closeNeeds')"
-    >
-      <Needs />
-    </Panel>
-
-    <Panel 
-      :isOpen="showCageStatus" 
-      title="Cage Status" 
-      @close="emit('closeCageStatus')"
-    >
-      <CageStatus />
-    </Panel>
-
-    <Panel 
-      :isOpen="showMarket" 
-      title="🐹 Guinea Pig Market" 
-      @close="emit('closeMarket')"
-    >
-      <Market />
-    </Panel>
+    <IconSidebarPanels 
+      :showInventory="showInventory"
+      :showGuineaPig="showGuineaPig"
+      :showNeeds="showNeeds"
+      :showCageStatus="showCageStatus"
+      :showMarket="showMarket"
+      @closeInventory="emit('closeInventory')"
+      @closeGuineaPig="emit('closeGuineaPig')"
+      @closeNeeds="emit('closeNeeds')"
+      @closeCageStatus="emit('closeCageStatus')"
+      @closeMarket="emit('closeMarket')"
+    />
   </div>
 </template>
